feat(modal-property): additional exit functionality
Adds a new property `escapable` to the modal component.
This property is set to true by default to improve
accesibility and user experience.

New ways to close the modal
  - pressing the escape key
  - clicking outside of the modal

It is sometimes important that a user interact with the content
in a modal before continuing. Maybe a terms of service, or age
verification. Setting the escapable property to false allows the
developer to close the modal only after the user meets certain
requirements. In this case, the escape key, outside exit clicks,
and the close button in the modal header are removed.

fix(accessibility): aria-label the close button for screenreaders

docs: adds an example and escapable info section to documentation