Blockquote component (#311)
* feat: Add FwbBlockquote component

This commit adds a new component called FwbBlockquote. It includes the template, script, and styles for rendering a blockquote element with customizable size, alignment, type, and citation. The component also utilizes the useBlockquoteClasses composable to generate the appropriate CSS classes based on the props.

Note: This commit message follows the established convention of starting with a type (feat) followed by a concise description of the changes.

* feat: Sample components for doc page

* feat: Blockquote doc page added

* fix: Update FwbBlockquote component

This commit updates the FwbBlockquote component in the typography/blockquote directory. The changes include:
- Fixing a typo in the component name from "FwbBlockquote" to "fwb-blockquote" to adhere to the naming convention.
- Adding missing spaces in the blockquote text to improve readability.
- Adding multiple blockquote examples with different sizes and alignments to the documentation page.

Note: This commit message follows the established convention of starting with a type (feat) followed by a concise description of the changes.

* fix: Update FwbBlockquote component name and fix typo

This commit updates the FwbBlockquote component in the typography/blockquote directory. The changes include:
- Fixing a typo in the component name from "FwbBlockquote" to "fwb-blockquote" to adhere to the naming convention.

Note: This commit message follows the established convention of starting with a type (chore) followed by a concise description of the changes.

* refactor: Remove deprecated FwbBlockquote component and related files

This commit removes the deprecated FwbBlockquote component and its related files:
- src/components/FwbBlockquote/FwbBlockquote.vue
- src/components/FwbBlockquote/composables/useBlockquoteClasses.ts
- src/components/FwbBlockquote/types.ts

The functionality of the FwbBlockquote component has been replaced by the new FwbBlockquote component located at src/components/Typography/FwbBlockquote.vue.

Note: This commit message follows the established convention of starting with a type (refactor) followed by a concise description of the changes.

* refactor: Apply lint rules to FwbBlockquote component and related files.

* fix: lint added

* refactor: Apply lint rules

---------

Co-authored-by: Oğuzhan Güç <oguzhan.guc@ayssoft.com>